services/mortgage_rates.py
"""
Mortgage rates service - placeholder for future API integration.
"""
import yaml
from pathlib import Path
from typing import Optional, Dict
from datetime import datetime


def get_data_path() -> Path:
    """Get the path to the data directory."""
    current_dir = Path(__file__).parent
    return current_dir.parent / "data"


def load_assumptions() -> dict:
    """Load assumptions from YAML file."""
    data_path = get_data_path() / "assumptions.yaml"
    
    try:
        with open(data_path, 'r') as f:
            return yaml.safe_load(f)
    except FileNotFoundError:
        # Return basic defaults
        return {
            "mortgage": {
                "typical_rate": 0.07,
                "typical_term_years": 30,
                "typical_down_payment": 0.20
            }
        }


def get_current_mortgage_rates(
    loan_type: str = "conventional",
    term_years: int = 30,
    location: Optional[str] = None
) -> Dict[str, float]:
    """
    Get current mortgage rates - placeholder for future API integration.
    
    Args:
        loan_type: Type of loan (conventional, fha, va, etc.)
        term_years: Loan term in years
        location: Location for regional rate adjustments
        
    Returns:
        Dictionary with rate information
    """
    assumptions = load_assumptions()
    base_rate = assumptions.get("mortgage", {}).get("typical_rate", 0.07)
    
    # Placeholder rate adjustments
    rate_adjustments = {
        "conventional_30": 0.0,
        "conventional_15": -0.003,  # 15-year typically 0.3% lower
        "fha": -0.002,              # FHA slightly lower
        "va": -0.0025,              # VA slightly lower
        "jumbo": 0.002,             # Jumbo slightly higher
    }
    
    loan_key = f"{loan_type}_{term_years}"
    adjustment = rate_adjustments.get(loan_key, rate_adjustments.get(loan_type, 0.0))
    
    current_rate = base_rate + adjustment
    
    # Regional adjustments (placeholder)
    regional_adjustment = 0.0
    if location and "NY" in location:
        regional_adjustment = 0.001  # Slightly higher in NYC
    elif location and "NJ" in location:
        regional_adjustment = 0.0005
    
    final_rate = current_rate + regional_adjustment
    
    return {
        "rate": final_rate,
        "base_rate": base_rate,
        "loan_type_adjustment": adjustment,
        "regional_adjustment": regional_adjustment,
        "last_updated": datetime.now().isoformat(),
        "source": "Static defaults - replace with API",
        "loan_type": loan_type,
        "term_years": term_years
    }


def get_rate_trends(days: int = 30) -> Dict[str, any]:
    """
    Get mortgage rate trends - placeholder for future implementation.
    
    Args:
        days: Number of days of historical data
        
    Returns:
        Dictionary with trend information
    """
    current_rates = get_current_mortgage_rates()
    
    # Placeholder trend data
    return {
        "current_rate": current_rates["rate"],
        "trend_direction": "stable",
        "change_30_days": 0.001,
        "change_90_days": -0.002,
        "volatility": "low",
        "forecast": "stable_to_slightly_higher",
        "data_source": "Placeholder - implement with real API"
    }


def get_points_options(base_rate: float, loan_amount: float) -> list[Dict[str, any]]:
    """
    Get mortgage points options - placeholder implementation.
    
    Args:
        base_rate: Base mortgage rate
        loan_amount: Loan amount
        
    Returns:
        List of points options with rates and costs
    """
    options = []
    
    # Generate options for 0, 0.5, 1, 1.5, 2 points
    for points in [0, 0.5, 1.0, 1.5, 2.0]:
        rate_reduction = points * 0.0025  # Assume 0.25% reduction per point
        new_rate = base_rate - rate_reduction
        cost = loan_amount * (points / 100)
        
        options.append({
            "points": points,
            "rate": new_rate,
            "cost": cost,
            "rate_reduction": rate_reduction,
            "break_even_months": int(cost / (loan_amount * rate_reduction / 12)) if rate_reduction > 0 else None
        })
    
    return options 