requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
plotly>=5.15.0
PyYAML>=6.0
numpy-financial>=1.0.0
ruff>=0.1.0 